requirements.txt
fastapi==0.115.8
python-multipart==0.0.20
uvicorn==0.34.0
pydantic==2.10.6
requests==2.32.3
pinecone-client[grpc]==5.0.1
ibmcloudant>=0.10.0
nltk==3.9.1
transformers==4.48.1
torch==2.5.1
dacite==1.8.1
beautifulsoup4==4.12.3
sentence-transformers==3.4.0
langchain-text-splitters==0.3.5
nbdev==2.3.34
rq==2.1.0
redis==5.2.1
pyjwt==2.10.1